Create busstationmanagementsystem.c
This is a console program written in C language. It has for loops so if you happen to  want to run it ensure  your IDE is in c99 mode. The purpose of this program is to add bus schedules, book a seat, cancel a booking, issue a ticket after a successful booking, view all cancelled tickets, view all issued tickets and view all the bus schedules added